sc2: Moved sc2 GUI to a separate file * Added and fixed many type annotations
